pyproject.toml
[tool.poetry]
name = "engine"
version = "0.1.0"
description = ""
authors = ["ddorn <diego.dorn@free.fr>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
pygame-ce = "^2.3.0"

[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
plotly = "^5.15.0"
numpy = "^1.25.2"
pandas = "^2.0.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
